cloud_master/file_management/validators/excel_file_serializers.py
from rest_framework.fields import FileField

from common.framework.serializer import BaseSerializer


class ExcelImportSerializer(BaseSerializer):
    file = FileField(required=True, allow_empty_file=False)
